Code and input dataset
K-means is the functional module, designed in a modular fashion. The different steps of the K-means clustering algorithm are separated into this file.

K-means is primarily the UI module, also modularly designed. A GUI prototype is designed using Mockplus to generate the UI code, and then the functional module is manually invoked to load data into the UI and implement data binding.

The overall project is a very simple front-end/back-end separation.